Flytter avstemmingsfunksjonene i OppdragKlient til AvstemmingKlient med lengre timeout
