+++ src/main/kotlin/no/nav/familie/ks/sak/kjerne/avstemming/AvstemmingService.kt
@@ -2,7 +2,7 @@ package no.nav.familie.ks.sak.kjerne.avstemming
 
 import no.nav.familie.kontrakter.felles.oppdrag.PerioderForBehandling
 import no.nav.familie.ks.sak.common.exception.Feil
-import no.nav.familie.ks.sak.integrasjon.oppdrag.OppdragKlient
+import no.nav.familie.ks.sak.integrasjon.oppdrag.AvstemmingKlient
 import no.nav.familie.ks.sak.integrasjon.secureLogger
 import no.nav.familie.ks.sak.kjerne.behandling.BehandlingService
 import no.nav.familie.ks.sak.kjerne.beregning.BeregningService
@@ -14,7 +14,7 @@ import java.util.UUID
 
 @Service
 class AvstemmingService(
-    private val oppdragKlient: OppdragKlient,
+    private val avstemmingKlient: AvstemmingKlient,
     private val behandlingService: BehandlingService,
     private val beregningService: BeregningService,
 ) {
@@ -23,7 +23,7 @@ class AvstemmingService(
         tom: LocalDateTime,
         avstemmingId: UUID?,
     ) {
-        oppdragKlient.sendGrensesnittavstemmingTilOppdrag(fom = fom, tom = tom, avstemmingId = avstemmingId)
+        avstemmingKlient.sendGrensesnittavstemmingTilOppdrag(fom = fom, tom = tom, avstemmingId = avstemmingId)
     }
 
     fun sendKonsistensavstemmingStartMelding(
@@ -31,7 +31,7 @@ class AvstemmingService(
         transaksjonsId: UUID,
     ) {
         logger.info("Utfører Konsistensavstemming: Sender start melding for transaksjonsId $transaksjonsId")
-        oppdragKlient.konsistensavstemOppdragStart(avstemmingstidspunkt, transaksjonsId)
+        avstemmingKlient.konsistensavstemOppdragStart(avstemmingstidspunkt, transaksjonsId)
     }
 
     fun sendKonsistensavstemmingData(
@@ -40,7 +40,7 @@ class AvstemmingService(
         transaksjonsId: UUID,
     ) {
         logger.info("Utfører Konsistensavstemming: Sender perioder for transaksjonsId $transaksjonsId")
-        oppdragKlient.konsistensavstemOppdragData(avstemmingstidspunkt, perioderTilAvstemming, transaksjonsId)
+        avstemmingKlient.konsistensavstemOppdragData(avstemmingstidspunkt, perioderTilAvstemming, transaksjonsId)
     }
 
     fun sendKonsistensavstemmingAvsluttMelding(
@@ -48,7 +48,7 @@ class AvstemmingService(
         transaksjonsId: UUID,
     ) {
         logger.info("Utfører Konsistensavstemming: Sender avslutt melding for transaksjonsId $transaksjonsId")
-        oppdragKlient.konsistensavstemOppdragAvslutt(avstemmingstidspunkt, transaksjonsId)
+        avstemmingKlient.konsistensavstemOppdragAvslutt(avstemmingstidspunkt, transaksjonsId)
     }
 
     fun hentDataForKonsistensavstemming(
